// src/types/modules/dnd/websiteContent.ts
export interface FieldEntry {
  label: string
  name: string
  dataFieldName: string
}

export interface WebsiteFormData {
  company_profile: string
  design_layout: string
  product_image: string
  product_features: string
  product_specifications: string
  contact_details: string
  regulatory_content: string
  logos: string
  mkgMaterial: string
  disclaimers: string
  legal_content: string
  marketing_elements: string
  navigation_functionality: string
}